@@ -1,6 +1,6 @@
 package com.tool.russ.view.navigation;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.widget.ImageView;
 import android.widget.TextView;
 
