@@ -1,7 +1,7 @@
 package com.tool.russ.view.navigation;
 
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 public interface NavigationListener {
 
