@@ -1,21 +1,24 @@
 package com.tool.russ.tool
 
 import android.annotation.SuppressLint
+import android.os.Bundle
 import android.os.Handler
 import android.os.Message
-import android.support.v7.app.AppCompatActivity
-import android.os.Bundle
-import android.support.v7.widget.LinearLayoutManager
-import android.support.v7.widget.RecyclerView
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.content.ContextCompat
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.tool.russ.view.ToolView
 import com.tool.russ.view.custom.ProgressNumber
 import com.tool.russ.view.dialog.TipDialog
-import com.tool.russ.view.refresh.RefreshListener
 import kotlinx.android.synthetic.main.activity_main.*
+
 
 class MainActivity : AppCompatActivity() {
     internal var progressNumber: ProgressNumber? = null
@@ -42,19 +45,29 @@ class MainActivity : AppCompatActivity() {
     @SuppressLint("HandlerLeak")
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        ToolView.init(this)
         setContentView(R.layout.activity_main)
 
 
 
-//        ToolView.init(this)
-//        val recyclerView = findViewById<RecyclerView>(R.id.list)
-//        recyclerView.adapter=object : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+
+
+
+
+
+
+//        list.adapter=object : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
 //            override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
 //                val view = LayoutInflater.from(parent.context).inflate(R.layout.item, parent, false)
 //                return object : RecyclerView.ViewHolder(view) {}
 //            }
 //
 //            override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
+//
+//                holder.itemView.post { Runnable {
+//                    Log.d(TAG,"第${holder.adapterPosition}个，width=${holder.itemView.width},height=${holder.itemView.height}")
+//                }.run() }
+//
 //
 //            }
 //
@@ -63,9 +76,20 @@ class MainActivity : AppCompatActivity() {
 //            }
 //        }
 //
-//        recyclerView.layoutManager = LinearLayoutManager(this)
+//        list.layoutManager = GridLayoutManager(this,2)
+//
+//        //添加Android自带的分割线
+//        val decoration=DividerItemDecoration(this, DividerItemDecoration.VERTICAL)
+//
+//        decoration.setDrawable(ContextCompat.getDrawable(this,R.drawable.divider)!!)
+//        list.addItemDecoration(decoration)
 //
 //
+//        val decoration1=DividerItemDecoration(this, DividerItemDecoration.HORIZONTAL)
+//
+//        decoration1.setDrawable(ContextCompat.getDrawable(this,R.drawable.divider)!!)
+//        list.addItemDecoration(decoration1)
+
 //        refresh.setListener(object:RefreshListener(){
 //            override fun refresh() {
 //                handler.sendEmptyMessageDelayed(2,3000)
